+++ lizard_progress/views/views.py
@@ -49,10 +49,8 @@ from lizard_progress.process_uploaded_file import make_uploaded_file_path
 logger = logging.getLogger(__name__)
 
 
-class ProjectsView(UiView):
-    """Displays a list of projects to choose from."""
-    template_name = "lizard_progress/progressbase.html"
-
+class ProjectsMixin(object):
+    """Helper functions for working with projects in views"""
     project_slug = None
     project = None
 
@@ -70,7 +68,7 @@ class ProjectsView(UiView):
             self.project = None
 
         return super(
-            ProjectsView, self).dispatch(request, *args, **kwargs)
+            ProjectsMixin, self).dispatch(request, *args, **kwargs)
 
     def projects(self):
         """Returns a list of projects the current user has access to."""
@@ -90,6 +88,11 @@ class ProjectsView(UiView):
         user = self.request.user
         userprofile = UserProfile.objects.get(user=user)
         return userprofile.profiletype == UserProfile.CONTRACTOR
+
+    def user_is_manager(self):
+        user = self.request.user
+        userprofile = UserProfile.objects.get(user=user)
+        return userprofile.profiletype == UserProfile.PROJECTMANAGER
 
     def project_home_url(self):
         if not self.project_slug:
@@ -121,79 +124,16 @@ class ProjectsView(UiView):
         return crumbs
 
 
-class View(AppView):
+class ProjectsView(ProjectsMixin, UiView):
+    """Displays a list of projects to choose from."""
+    template_name = "lizard_progress/progressbase.html"
+
+
+class View(ProjectsMixin, AppView):
     """The app's root, shows a choice of projects, or a choice of
     dashboard / upload / map layer pages if a project is chosen."""
 
-    project_slug = None
-    project = None
     template_name = 'lizard_progress/home.html'
-
-    def dispatch(self, request, *args, **kwargs):
-
-        self.project_slug = kwargs.get('project_slug', None)
-        if self.project_slug is None:
-            self.project = Project.objects.all()[0]
-            self.project_slug = self.project.slug
-        else:
-            self.project = get_object_or_404(Project, slug=self.project_slug)
-
-            if has_access(request.user, self.project):
-                self.has_full_access = all(
-                    has_access(request.user, self.project, contractor)
-                    for contractor in self.project.contractor_set.all())
-            else:
-                raise PermissionDenied()
-
-        return super(View, self).dispatch(request, *args, **kwargs)
-
-    def upload_home_url(self):
-        """Returns URL to this project's Upload view"""
-        return reverse('lizard_progress_uploadhomeview',
-                       kwargs={'project_slug': self.project_slug})
-
-    def download_home_url(self):
-        """Returns URL to this project's Download view"""
-        return reverse('lizard_progress_downloadhomeview',
-                       kwargs={'project_slug': self.project_slug})
-
-    def dashboard_url(self):
-        """Returns URL to this project's Dashboard view"""
-        return reverse('lizard_progress_dashboardview',
-                       kwargs={'project_slug': self.project_slug})
-
-    def map_url(self):
-        """Returns URL to this project's Map view"""
-        return reverse('lizard_progress_mapview',
-                       kwargs={'project_slug': self.project_slug})
-
-    def comparison_url(self):
-        """Returns URL to this project's Comparison view"""
-        return reverse('lizard_progress_comparisonview',
-                       kwargs={'project_slug': self.project_slug})
-
-    def admin_url(self):
-        """Returns URL to this project's Comparison view"""
-        return reverse('lizard_progress_admin')
-
-    @property
-    def breadcrumbs(self):
-        """Returns a list of breadcrumbs to this project."""
-        crumbs = [
-            Action(
-                description="Home",
-                name="Home",
-                url="/")]
-
-        if self.project:
-            crumbs.append(
-                Action(
-                    description=self.project.name,
-                    name=self.project.name,
-                    url=reverse('lizard_progress_view',
-                                kwargs={'project_slug': self.project_slug})))
-
-        return crumbs
 
 
 class MapView(View):
@@ -245,6 +185,22 @@ class MapView(View):
             })
 
         return layers
+
+    @property
+    def breadcrumbs(self):
+        """Breadcrumbs for this page."""
+
+        crumbs = super(MapView, self).breadcrumbs
+
+        crumbs.append(
+            Action(
+                name="Kaartlagen",
+                description=("De kaartlagen van {project} in Lizard"
+                             .format(project=self.project.name)),
+                url=reverse('lizard_progress_mapview',
+                            kwargs={'project_slug': self.project_slug})))
+
+        return crumbs
 
 
 class ComparisonView(View):
@@ -480,23 +436,27 @@ class ComparisonPopupView(View):
         return htmls
 
 
-class DashboardView(ProjectsView, View):
+class DashboardView(ProjectsView):
     """Show the dashboard page. The page offers a popup per area per
     contractor, if the user has access, and also downloads to CSV
     files."""
 
     template_name = 'lizard_progress/dashboard.html'
 
-    def crumbs(self):
+    @property
+    def breadcrumbs(self):
         """Breadcrumbs for this page."""
 
-        crumbs = super(DashboardView, self).crumbs()
+        crumbs = super(DashboardView, self).breadcrumbs
 
-        crumbs.append({
-                'url': self.dashboard_url(),
-                'description': 'Dashboard',
-                'title': '%s dashboard' % (self.project.name,)
-                })
+        crumbs.append(
+            Action(
+                name="Dashboard",
+                description=("{project} dashboard"
+                             .format(project=self.project.name)),
+                url=reverse(
+                    'lizard_progress_dashboardview',
+                    kwargs={'project_slug': self.project_slug})))
 
         return crumbs
 
@@ -516,24 +476,6 @@ class DashboardView(ProjectsView, View):
                                     'area_slug': area.slug})))
 
         return areas
-
-    def csv(self):
-        """Links to CSV downloads. One per contractor."""
-
-        csvs = []
-
-        for contractor in Contractor.objects.filter(project=self.project):
-            if has_access(self.request.user, self.project, contractor):
-                url = reverse(
-                    'lizard_progress_dashboardcsvview',
-                    kwargs={
-                        'project_slug': self.project_slug,
-                        'contractor_slug': contractor.slug,
-                        })
-
-                csvs.append((contractor, url))
-
-        return csvs
 
 
 class DashboardAreaView(View):
@@ -713,7 +655,11 @@ def dashboard_graph(request, project_slug, contractor_slug,
         raise PermissionDenied()
 
     fig = ScreenFigure(600, 350)  # in pixels
-    fig.text(0.5, 0.85, 'Uitgevoerde werkzaamheden', fontsize=14, ha='center')
+    fig.text(
+        0.5, 0.85,
+        ('Uitgevoerde werkzaamheden {contractor}'
+         .format(contractor=contractor.organization.name)),
+        fontsize=14, ha='center')
     fig.subplots_adjust(left=0.05, right=0.95)  # smaller margins
     y_title = -0.2  # a bit lower